EDU-416 Use dedicated icon for user item edited activity
